backend/requirements.txt
Flask==2.3.3
python-dotenv==1.0.0
# Database
flask-sqlalchemy==3.1.1
flask-migrate==4.0.5
flask-cors==4.0.0
python-jose==3.3.0
passlib==1.7.4
requests==2.31.0

# Web Scraping
beautifulsoup4==4.14.2
lxml==6.0.2
gunicorn>=21.2.0

# Utils
PySocks==1.7.1
pytweening==1.2.0
schedule==1.2.2

# Note: The following packages are macOS-specific and have been removed for Docker compatibility:
# - pyobjc-framework-Cocoa
# - pyobjc-framework-Quartz
# - pyperclip
# - PyRect
# - PyScreeze
# - rubicon-objc
selenium==4.36.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.8
trio==0.31.0
attrs==25.4.0
certifi==2025.10.5
charset-normalizer==3.4.4
h11==0.16.0
idna==3.11
MouseInfo==0.1.3
outcome==1.3.0.post0
PyAutoGUI==0.9.54
PyGetWindow==0.0.9
PyMsgBox==2.0.1
typing_extensions==4.15.0
urllib3==2.5.0
websocket-client==1.9.0
wsproto==1.2.0
